+++ Sources/LearningImpl/src/bagaturchess/deeplearning/impl/eval/pst_and_allfeatures/NeuralNetworkEvaluator.java
@@ -38,19 +38,19 @@ public class NeuralNetworkEvaluator extends BaseEvaluator {
 	
 	
 	@Override
-	protected double phase1() {
+	protected int phase1() {
 		
 		NeuralNetworkUtils.clearInputsArray(inputs);
 		NeuralNetworkUtils_PST_And_AllFeatures.fillInputs(network, inputs, bitboard, filler);
 		NeuralNetworkUtils.calculate(network);
 		double actualWhitePlayerEval = NeuralNetworkUtils.getOutput(network);
 		
-		return actualWhitePlayerEval;
+		return (int) actualWhitePlayerEval;
 	}
 	
 	
 	@Override
-	protected double phase2() {
+	protected int phase2() {
 
 		int eval = 0;
 		
@@ -59,7 +59,7 @@ public class NeuralNetworkEvaluator extends BaseEvaluator {
 	
 	
 	@Override
-	protected double phase3() {
+	protected int phase3() {
 		
 		int eval = 0;
 				
@@ -68,7 +68,7 @@ public class NeuralNetworkEvaluator extends BaseEvaluator {
 	
 	
 	@Override
-	protected double phase4() {
+	protected int phase4() {
 		
 		int eval = 0;
 		
@@ -77,7 +77,7 @@ public class NeuralNetworkEvaluator extends BaseEvaluator {
 	
 	
 	@Override
-	protected double phase5() {
+	protected int phase5() {
 		
 		int eval = 0;
