+++ Sources/LearningImpl/src/bagaturchess/deeplearning/impl_nnue/eval/NeuralNetworkEvaluator.java
@@ -49,7 +49,7 @@ public class NeuralNetworkEvaluator extends BaseEvaluator {
 	
 	
 	@Override
-	protected double phase1() {
+	protected int phase1() {
 		
 		Tensor tensor = NNUE_Constants.createInput(bitboard, inputs_3d);
 		
@@ -62,12 +62,12 @@ public class NeuralNetworkEvaluator extends BaseEvaluator {
 
 		actualWhitePlayerEval = activation_function.SIGMOID.getx(actualWhitePlayerEval);
 		
-		return actualWhitePlayerEval;
+		return (int) actualWhitePlayerEval;
 	}
 	
 	
 	@Override
-	protected double phase2() {
+	protected int phase2() {
 
 		int eval = 0;
 		
@@ -76,7 +76,7 @@ public class NeuralNetworkEvaluator extends BaseEvaluator {
 	
 	
 	@Override
-	protected double phase3() {
+	protected int phase3() {
 		
 		int eval = 0;
 				
@@ -85,7 +85,7 @@ public class NeuralNetworkEvaluator extends BaseEvaluator {
 	
 	
 	@Override
-	protected double phase4() {
+	protected int phase4() {
 		
 		int eval = 0;
 		
@@ -94,7 +94,7 @@ public class NeuralNetworkEvaluator extends BaseEvaluator {
 	
 	
 	@Override
-	protected double phase5() {
+	protected int phase5() {
 		
 		int eval = 0;
