+++ Sources/LearningImpl/src/bagaturchess/learning/goldmiddle/impl/cfg/bagatur/eval/BagaturEvaluator.java
@@ -88,7 +88,7 @@ public class BagaturEvaluator extends BaseEvaluator implements FeatureWeights {
 	
 	
 	@Override
-	protected double phase1() {
+	protected int phase1() {
 		int eval = 0;
 		
 		evalInfo.clear_short();
@@ -109,7 +109,7 @@ public class BagaturEvaluator extends BaseEvaluator implements FeatureWeights {
 	
 	
 	@Override
-	protected double phase2() {
+	protected int phase2() {
 
 		int eval = 0;
 		
@@ -136,7 +136,7 @@ public class BagaturEvaluator extends BaseEvaluator implements FeatureWeights {
 	
 	
 	@Override
-	protected double phase3() {
+	protected int phase3() {
 		
 		int eval = 0;
 				
@@ -166,7 +166,7 @@ public class BagaturEvaluator extends BaseEvaluator implements FeatureWeights {
 	
 	
 	@Override
-	protected double phase4() {
+	protected int phase4() {
 		
 		int eval = 0;
 		
@@ -201,7 +201,7 @@ public class BagaturEvaluator extends BaseEvaluator implements FeatureWeights {
 	
 	
 	@Override
-	protected double phase5() {
+	protected int phase5() {
 		
 		int eval = 0;
