+++ Sources/LearningImpl/src/bagaturchess/learning/goldmiddle/impl/cfg/bagatur_allfeatures/eval/BagaturEvaluator.java
@@ -83,7 +83,7 @@ public class BagaturEvaluator extends BaseEvaluator implements FeatureWeights {
 	
 	
 	@Override
-	protected double phase1() {
+	protected int phase1() {
 		int eval = 0;
 		
 		evalInfo.clear_short();
@@ -100,7 +100,7 @@ public class BagaturEvaluator extends BaseEvaluator implements FeatureWeights {
 	
 	
 	@Override
-	protected double phase2() {
+	protected int phase2() {
 
 		int eval = 0;
 		
@@ -121,7 +121,7 @@ public class BagaturEvaluator extends BaseEvaluator implements FeatureWeights {
 	
 	
 	@Override
-	protected double phase3() {
+	protected int phase3() {
 		
 		int eval = 0;
 				
@@ -145,7 +145,7 @@ public class BagaturEvaluator extends BaseEvaluator implements FeatureWeights {
 	
 	
 	@Override
-	protected double phase4() {
+	protected int phase4() {
 		
 		int eval = 0;
 		
@@ -172,7 +172,7 @@ public class BagaturEvaluator extends BaseEvaluator implements FeatureWeights {
 	
 	
 	@Override
-	protected double phase5() {
+	protected int phase5() {
 		
 		int eval = 0;
